+++ addon/components/paper-icon.js
@@ -1,8 +1,16 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
 import ColorMixin from 'ember-paper/mixins/color-mixin';
 
 const { Component, computed, String: Str } = Ember;
 
+/**
+ * @class PaperIcon
+ * @extends Ember.Component
+ * @uses ColorMixin
+ */
 let PaperIconComponent = Component.extend(ColorMixin, {
   tagName: 'md-icon',
   classNames: ['paper-icon', 'md-font', 'material-icons', 'md-default-theme'],