+++ addon/components/paper-list.js
@@ -1,6 +1,15 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
 
-export default Ember.Component.extend({
+const { Component } = Ember;
+
+/**
+ * @class PaperList
+ * @extends Ember.Component
+ */
+export default Component.extend({
   tagName: 'md-list',
   classNames: ['paper-list', 'md-default-theme']
 });