+++ addon/components/paper-menu-item.js
@@ -1,8 +1,16 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
 import ChildMixin from 'ember-paper/mixins/child-mixin';
 
 const { Component } = Ember;
 
+/**
+ * @class PaperMenuItem
+ * @extends Ember.Component
+ * @uses ChildMixin
+ */
 export default Component.extend(ChildMixin, {
   tagName: 'md-menu-item',
   disabled: false,