+++ addon/components/paper-option.js
@@ -1,9 +1,17 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
 import PaperMenuItem from './paper-menu-item';
 import RippleMixin from '../mixins/ripple-mixin';
 
 const { computed } = Ember;
 
+/**
+ * @class PaperOption
+ * @extends PaperMenuItem
+ * @uses RippleMixin
+ */
 export default PaperMenuItem.extend(RippleMixin, {
   tagName: 'md-option',
   attributeBindings: ['aria-selected','aria-disabled','aria-current','data-option-index','role','selected','tabindex'],