+++ addon/components/paper-radio.js
@@ -1,10 +1,21 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
 import BaseFocusable from './base-focusable';
 import RippleMixin from 'ember-paper/mixins/ripple-mixin';
 import ColorMixin from 'ember-paper/mixins/color-mixin';
 import ChildMixin from 'ember-paper/mixins/child-mixin';
+
 const { computed, assert } = Ember;
 
+/**
+ * @class PaperRadio
+ * @extends BaseFocusable
+ * @uses ColorMixin
+ * @uses RippleMixin
+ * @uses ChildMixin
+ */
 export default BaseFocusable.extend(RippleMixin, ColorMixin, ChildMixin, {
   tagName: 'md-radio-button',
   classNames: ['md-default-theme'],