+++ addon/components/paper-select-trigger.js
@@ -1,8 +1,15 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
 import TriggerComponent from 'ember-power-select/components/power-select/trigger';
 
 const { computed } = Ember;
 
+/**
+ * @class PaperSelectTrigger
+ * @extends Ember.Component
+ */
 export default TriggerComponent.extend({
   tagName: 'md-select-value',
   classNames: ['md-select-value'],