+++ addon/components/paper-toolbar-tools.js
@@ -1,6 +1,14 @@
+/**
+ * @module ember-paper
+ */
 import Ember from 'ember';
+
 const { Component } = Ember;
 
+/**
+ * @class PaperToolbarTools
+ * @extends Ember.Component
+ */
 export default Component.extend({
   classNames: ['md-toolbar-tools']
 });