+++ app/services/sniffer.js
@@ -1,3 +1,4 @@
+/* globals FastBoot */
 import Ember from 'ember';
 
 let isString = function(value) {
@@ -16,17 +17,28 @@ export default Ember.Service.extend({
   vendorPrefix: '',
   transitions: false,
   animations: false,
-  document,
-  window,
+  _document: null,
+  _window: null,
 
   android: Ember.computed('', function() {
-    return toInt((/android (\d+)/.exec(lowercase((this.get('window').navigator || {}).userAgent)) || [])[1]);
+    return toInt((/android (\d+)/.exec(lowercase((this.get('_window').navigator || {}).userAgent)) || [])[1]);
   }),
 
   init() {
     this._super(...arguments);
+    if (typeof FastBoot !== 'undefined') {
+      return;
+    }
 
-    let bodyStyle = this.get('document').body && this.get('document').body.style;
+    let _document = document;
+    let _window = window;
+
+    this.setProperties({
+      _document,
+      _window
+    });
+
+    let bodyStyle = _document.body && _document.body.style;
     let vendorPrefix;
     let vendorRegex = /^(Moz|webkit|ms)(?=[A-Z])/;