+++ tests/dummy/snippets/install-0.2.sh
@@ -1,1 +1,1 @@
-$ ember install ember-paper
+$ ember install ember-paper@^0.2.0